% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalfunctions.R
\name{KaplanMeier}
\alias{KaplanMeier}
\title{Kaplan Meier Product Limit Estimator}
\usage{
KaplanMeier(x, rs)
}
\arguments{
\item{x}{The value to pass to the Kaplan Meier function}

\item{rs}{is the output from RiskSet(...)}
}
\value{
The estimate of the survival function at a given value of x
}
\description{
Survival function estimate using Kaplan Meier method
}
\examples{
x <- c(1.0,1.3,1.5,1.5,2.1,2.1,2.1,2.8)
rs <- RiskSet(x)
plot.stepfun(ecdf(x), verticals=TRUE, pch='',
             ylab="F(x)", main="Empirical CDF")
plot(function(y) NelsonAalen(y,rs), 0, 3,
     type='s', ylab="H(x)", main="Nelson-Aalen estimate of H(x)")
}

