% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalfunctions.R
\name{NelsonAalen}
\alias{NelsonAalen}
\title{Nelson Aalen Estimator}
\usage{
NelsonAalen(x, rs, survival = FALSE)
}
\arguments{
\item{x}{The value to pass to the Nelson Aalen function}

\item{rs}{is the output from RiskSet(...)}

\item{survival}{If TRUE return estimate of survival function, otherwise return estimate of cumulative hazard function}
}
\value{
The estimate  of the cumulative hazard function at a given value of x
}
\description{
Cumulative hazard function estimate using Nelson Aalen method
}
\details{
To get estimate of survival function take exp(-H(x)) where H(x) is the cumulative hazard rate
}
\examples{
x <- c(1.0,1.3,1.5,1.5,2.1,2.1,2.1,2.8)
rs <- RiskSet(x)
plot(function(y) NelsonAalen(y,rs), 0, 3,
     type='s', ylab="H(x)", main="Nelson-Aalen estimate of H(x)")
}

