% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalfunctions.R
\name{ogive}
\alias{ogive}
\title{Ogive}
\usage{
ogive(x, cdf, breaks)
}
\arguments{
\item{x}{The value to pass to the ogive function}

\item{cdf}{A cdf function to use (use ecdf on the data)}

\item{breaks}{The breaks for the ogive function}
}
\value{
The ogive value (between 0 and 1) for a given x and cdf function.
}
\description{
Computes the ogive value for a given x
}
\details{
The cdf is computed before the ogive function is called so
         it doesn't have to be recomputed every time ogive(...) is called.
}
\examples{
x <- c(8,9,13,15,16,22,23,27,29,35)
c <- c(0,10,15,20,30,40)
f <- ecdf(x)
plot(function(y) ogive(y,f,c), 0, 50, main="Ogive", ylab="F(x)")
}

