% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalfunctions.R
\name{RiskSet}
\alias{RiskSet}
\title{Risk Set}
\usage{
RiskSet(x, d = NA, u = NA)
}
\arguments{
\item{x}{The data}

\item{d}{The left truncation points (optional)}

\item{u}{The right censoring points (optional)}
}
\value{
A data frame containing the unique values of x (y),
        the counts for each unique value (s), and the risk
        set (r)
}
\description{
Computes the risk set for a given vector x
}
\details{
If d and u are provided, they should be the same lenggth as
x. Entries of d, u and x can be set to NA where applicable.
}
\examples{
x <- c(1.0,1.3,1.5,1.5,2.1,2.1,2.1,2.8)
rs <- RiskSet(x)
}

